+++ tests/dotnet-e2e/outgoing/Program.cs
@@ -1,0 +1,8 @@
+﻿using Flurl.Http;
+
+var response = await "http://www.google.com/robots.txt".GetAsync();
+
+if (response.StatusCode > 299)
+    throw new Exception("Response failed with status code");
+
+Console.WriteLine(await response.GetStringAsync());